+++ test/services/FeedProcessor.test.ts
@@ -17,16 +17,18 @@ describe('FeedProcessor', () => {
   let settings: Settings;
 
   const mockedFetchers = {
-    CryptoCompareHistoHourFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CryptoCompareHistoHourFetcher>,
-    CryptoCompareHistoDayFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CryptoCompareHistoDayFetcher>,
-    CryptoComparePriceFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CryptoComparePriceFetcher>,
-    GVolImpliedVolatilityFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.GVolImpliedVolatilityFetcher>,
-    PolygonIOPriceFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.PolygonIOPriceFetcher>,
-    CryptoComparePriceWSFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CryptoComparePriceWSFetcher>,
-    IEXEnergyFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.IEXEnergyFetcher>,
-    CoingeckoPriceFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CoingeckoPriceFetcher>,
-    CoinmarketcapPriceFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.CoinmarketcapPriceFetcher>,
-    BEACPIAverageFetcher: (null as unknown) as sinon.SinonStubbedInstance<fetchers.BEACPIAverageFetcher>,
+    CryptoCompareHistoHourFetcher:
+      null as unknown as sinon.SinonStubbedInstance<fetchers.CryptoCompareHistoHourFetcher>,
+    CryptoCompareHistoDayFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.CryptoCompareHistoDayFetcher>,
+    CryptoComparePriceMultiFetcher:
+      null as unknown as sinon.SinonStubbedInstance<fetchers.CryptoComparePriceMultiFetcher>,
+    GVolImpliedVolatilityFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.GVolImpliedVolatilityFetcher>,
+    PolygonIOPriceFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.PolygonIOPriceFetcher>,
+    CryptoComparePriceWSFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.CryptoComparePriceWSFetcher>,
+    IEXEnergyFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.IEXEnergyFetcher>,
+    CoingeckoPriceFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.CoingeckoPriceFetcher>,
+    CoinmarketcapPriceFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.CoinmarketcapPriceFetcher>,
+    BEACPIAverageFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.BEACPIAverageFetcher>,
   };
 
   let feedProcessor: FeedProcessor;
@@ -34,7 +36,7 @@ describe('FeedProcessor', () => {
   beforeEach(async () => {
     mockedFetchers.CryptoCompareHistoHourFetcher = sinon.createStubInstance(fetchers.CryptoCompareHistoHourFetcher);
     mockedFetchers.CryptoCompareHistoDayFetcher = sinon.createStubInstance(fetchers.CryptoCompareHistoDayFetcher);
-    mockedFetchers.CryptoComparePriceFetcher = sinon.createStubInstance(fetchers.CryptoComparePriceFetcher);
+    mockedFetchers.CryptoComparePriceMultiFetcher = sinon.createStubInstance(fetchers.CryptoComparePriceMultiFetcher);
     mockedFetchers.GVolImpliedVolatilityFetcher = sinon.createStubInstance(fetchers.GVolImpliedVolatilityFetcher);
     mockedFetchers.PolygonIOPriceFetcher = sinon.createStubInstance(fetchers.PolygonIOPriceFetcher);
     mockedFetchers.CryptoComparePriceWSFetcher = sinon.createStubInstance(fetchers.CryptoComparePriceWSFetcher);
@@ -59,7 +61,9 @@ describe('FeedProcessor', () => {
     container
       .bind(fetchers.CryptoCompareHistoDayFetcher)
       .toConstantValue(mockedFetchers.CryptoCompareHistoDayFetcher as any);
-    container.bind(fetchers.CryptoComparePriceFetcher).toConstantValue(mockedFetchers.CryptoComparePriceFetcher as any);
+    container
+      .bind(fetchers.CryptoComparePriceMultiFetcher)
+      .toConstantValue(mockedFetchers.CryptoComparePriceMultiFetcher as any);
     container
       .bind(fetchers.GVolImpliedVolatilityFetcher)
       .toConstantValue(mockedFetchers.GVolImpliedVolatilityFetcher as any);